Move errorkind to consumererror
