+++ cmd/occollector/app/sender/jaeger_thrift_tchannel_sender.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/open-telemetry/opentelemetry-service/consumer"
 	"github.com/open-telemetry/opentelemetry-service/consumer/consumerdata"
-	"github.com/open-telemetry/opentelemetry-service/errors/errorkind"
+	"github.com/open-telemetry/opentelemetry-service/consumer/consumererror"
 	jaegertranslator "github.com/open-telemetry/opentelemetry-service/translator/trace/jaeger"
 )
 
@@ -51,7 +51,7 @@ func (s *JaegerThriftTChannelSender) ConsumeTraceData(ctx context.Context, td co
 	// TODO: (@pjanotti) In case of failure the translation to Jaeger Thrift is going to be remade, cache it somehow.
 	tBatch, err := jaegertranslator.OCProtoToJaegerThrift(td)
 	if err != nil {
-		return errorkind.Permanent(err)
+		return consumererror.Permanent(err)
 	}
 
 	if err := s.reporter.EmitBatch(tBatch); err != nil {
